token logic changed and improved & preparing LoginForm component for form validaton
